Relocate fido command
