+++ cmd/fido/cmd/json.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/poka-yoke/spaceflight/mcc/fido/fido"
+	"github.com/poka-yoke/spaceflight/pkg/fido"
 )
 
 // jsonCmd represents the json command
