+++ cmd/fido/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/poka-yoke/spaceflight/cmd/fido/cmd"
+
+func main() {
+	cmd.Execute()
+}
