--- mcc/fido/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/poka-yoke/spaceflight/mcc/fido/cmd"
-
-func main() {
-	cmd.Execute()
-}
